Rename 'components/MonitoringItemsOverviewTableValueWithTranslatableUnit.vue' to 'components/index/CardsMonitoring/MonitoringItemsOverview/ValueWithTranslatableUnit.vue'
