+++ components/index/CardsMonitoring/MonitoringItemsOverview/MedicalSystem.vue
@@ -13,7 +13,7 @@
             <li :class="[$style.box]">
               <div :class="$style.content">
                 <span>{{ $t('陽性率') }}</span>
-                <monitoring-items-overview-table-value-with-translatable-unit
+                <value-with-translatable-unit
                   :value="items['(4)PCR・抗原検査（陽性率）'].value"
                   :unit="items['(4)PCR・抗原検査（陽性率）'].unit"
                 />
@@ -22,7 +22,7 @@
             <li :class="[$style.box]">
               <div :class="$style.content">
                 <span>{{ $t('検査人数') }}</span>
-                <monitoring-items-overview-table-value-with-translatable-unit
+                <value-with-translatable-unit
                   :value="items['(4)PCR・抗原検査（検査人数）'].value"
                   :unit="items['(4)PCR・抗原検査（検査人数）'].unit"
                 />
@@ -40,7 +40,7 @@
         <li :class="[$style.box]">
           <div :class="$style.content">
             <span>{{ $t('(5)救急医療の東京ルールの適用件数') }}</span>
-            <monitoring-items-overview-table-value-with-translatable-unit
+            <value-with-translatable-unit
               :value="items['(5)救急医療の東京ルールの適用件数'].value"
               :unit="items['(5)救急医療の東京ルールの適用件数'].unit"
             />
@@ -49,7 +49,7 @@
         <li :class="[$style.box, $style.parent]">
           <div :class="$style.content">
             <span>{{ $t('(6)入院患者数') }}</span>
-            <monitoring-items-overview-table-value-with-translatable-unit
+            <value-with-translatable-unit
               :value="items['(6)入院患者数'].value"
               :unit="items['(6)入院患者数'].unit"
             />
@@ -66,7 +66,7 @@
         <li :class="[$style.box, $style.parent]">
           <div :class="$style.content">
             <span>{{ $t('(7)重症患者数') }}</span>
-            <monitoring-items-overview-table-value-with-translatable-unit
+            <value-with-translatable-unit
               :value="items['(7)重症患者数'].value"
               :unit="items['(7)重症患者数'].unit"
             />
@@ -88,12 +88,12 @@
 <script lang="ts">
 import Vue, { PropType } from 'vue'
 
-import MonitoringItemsOverviewTableValueWithTranslatableUnit from '@/components/MonitoringItemsOverviewTableValueWithTranslatableUnit.vue'
+import ValueWithTranslatableUnit from '@/components/index/CardsMonitoring/MonitoringItemsOverview/ValueWithTranslatableUnit.vue'
 import { MonitoringItems } from '@/utils/formatMonitoringItems'
 
 export default Vue.extend({
   components: {
-    MonitoringItemsOverviewTableValueWithTranslatableUnit,
+    ValueWithTranslatableUnit,
   },
   props: {
     items: {
